add mushroom enemy with behavior and integrate into level

+++ scripts/mushroom.gd
@@ -27,17 +27,19 @@ func _physics_process(delta: float) -> void:
 	
 	move_and_slide()
 
-#func _on_kill_zone_body_entered(body: Node2D) -> void:
-	#if body.is_in_group("Player"):
-		#body.move_local_y(-20)
-		##set_physics_process(false)
-		##$KillZone/CollisionShape2D.disabled = true
-		#skin.play("hit")
-		#await skin.animation_finished
-		#queue_free() #Entfernt die Frücht komplett
 
 
-func _on_hurt_player_zone_body_entered(body: Node2D) -> void:
+func _on_area_2d_body_entered(body: Node2D) -> void:
 	if body.is_in_group("Player"):
-		await body.delete()
-		body.reset(GameManager.LevelStartZone)
+		var y_delta = position.y - body.position.y
+		# Differenz zwischen Gegner- und Spieler-Position.
+		# Wenn Spieler von der Seite berührt, ist die Differenz klein,
+		# von oben ist die Differenz groß:
+		if y_delta < 10: #Ge
+			await body.delete()
+			body.reset(GameManager.LevelStartZone)
+		else: #Gegner stirbt
+			skin.play("hit")
+			body.bounce(0.8)
+			await skin.animation_finished
+			queue_free() #Entfernt Gegner aus dem Spiel
